ReadV: refactoring
Previous implementation of readv is the following: every readv request
is executed as a sequence of synchronous read of indivitual chunks. It
has poor performance, which affect significanlty vo operatoins,
the LHCb ones.

To improve the situation we introduce several changes:
1. We use local buffering. The idea is to read huge pice of file
and then extract requested chunks from this piece.
2. We do not use rados striper for the reads, but read data directly
from ceph objects. This seem to improve performance significantly.
3. If a piece of file that should be read occupies multiple ceph
objects, reads are executed asynchronously.