feat: add file locking to Xrootd delete and write operations
This might be useful to prevent data loss/corruption when we have
simultaneous transfers of the same file.

We do the locking in the following way: locks are added on a separate
object with the name <file_name>.<lock_name>, where <lock_name> is
different from file's object numbers (it is basically a text string).
On lock release, this object is removed. We use this awkward tecnique
instead of locking the first object for the following reason:
radosstriper does not like when he tries to create a new file, and
its first object already exists.